requirements.txt
torch
torchvision
torchaudio
diffusers
transformers
accelerate
safetensors
streamlit
pillow